feat: Update webapp dependencies
Update the dependencies of the static web app.
As react-scripts is not actively maintained, a migration from react-scripts (with webpack) to Vite was required to ensure we can use the latest versions of the dependencies.
The managed function used in the SWA was not much updated as there is no support for Fuctions v4 and Node20.

noissue